requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
paramiko==3.3.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
psutil==5.9.6
bcrypt==4.1.1
python-dotenv==1.0.0
asyncssh==2.14.1
pydantic==2.5.0
pydantic-settings==2.1.0